requirements.txt
pandas
numpy
matplotlib
sklearn
gunicorn
flask
pyyaml
scipy